breaking up the logic
